feat(hypervisor): Pipeline module extraction + surface interaction kit — zero drift (#56)
Pipeline becomes the first real app module; the serve monolith shrinks by the
whole pipeline block (~330 lines) plus the shared global rail (~60 lines).

Module: surfaces/pipeline/index.mjs under the surface contract { meta, load(ctx),
render(model, ctx), actions } — loaders and renderer moved VERBATIM (local
CX_ESC = escHtml alias keeps the block byte-identical). The registry imports and
binds the module itself (bindSurface("pipeline", pipelineModule)); the serve's
inline binding is gone and its dispatch speaks the module contract. Route stays
/__ioi/pipeline at the same chain position.

Shared chrome: surfaces/chrome.mjs — ioiGlobalRailHtml + IOI_GRAIL_CSS moved
verbatim (pixel-frozen; every certified port renders it); serve imports it, the
other 13 rail call sites unchanged.

Interaction kit: surfaces/kit.mjs — deliberately UNWIRED primitives for the node-
selection PR: escHtml (THE canonical escaper, serve now aliases it; two stale
local escaper copies deduped), parseSelection/selectionQuery (URL = selection's
single source of truth, stable key order), inspectorShell, trayShell,
disabledCommand (visible + names its reason), proofLink, semanticMask.

Zero drift proven: /__ioi/pipeline HTML byte-identical before/after extraction
(captured live, diffed twice). Pipeline pixel cert artifact untouched.

Verifiers: NEW verify-hypervisor-surface-modules.mjs 22/22 (contract, registry/
matrix/meta agreement, identity binding, offline dead-daemon render keeps the
certified landmarks, extraction hygiene — monolith no longer defines the moved
code, single escaper definition — and kit unit checks). Done bar: pipeline 31/31,
runtime-safety 20/20, reset 23/23, pixel harness 37/37, approvals 23/23, models
17/17, home-surface 56/56, matrix --check clean, no NUL, no Rust.
